refactor: 로그인 로직 클래스 병합 - SocialService -> AuthService 클래스 명 변경 - AppleLoginService 제거, AuthService로 통합
